fix: Don't set core.excludesFile
I thought SimpleGit.addConfig() would temporarily alter the current git
config, but it turns out that it permanently modifies the local git
config. This is highly undesirable, as a leftover .gitignore inside the
build/ dir would affect the entire repository.

Solve this by NOT using SimpleGit.addConfig() to apply a .gitignore.
Instead, copy release.gitignore into the build/dist/ dir AND rename it
to .gitignore, so that it is also copied back to project root (and
therefore apply to the current commit) while updating the release
branch.