fix: Use toInt() + toFamiliar() instead of Familiar.get()
toFamiliar() does not recognize familiar IDs that are JavaScript
integers, because all JavaScript numbers are inherently floating-point.
It seems that using toInt() generates "proper" integers that
toFamiliar() can understand.

Using toFamiliar() is more resilient than Familiar.get() since it will
not crash the runtime if an unknown ID is encountered.